PROJECT_ETL/event_producers/requirements.txt
kafka-python
pyarrow
pandas